+++ src/Controller/EvaluationController.php
@@ -23,9 +23,22 @@ class EvaluationController extends AbstractController
      */
     public function index(EvaluationRepository $evaluationRepository): Response
     {
-        return $this->render('evaluation/index.html.twig', [
-            'evaluations' => $evaluationRepository->findAll(),
-        ]);
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_RH')) {
+            return $this->render('evaluation/index.html.twig', [
+                'evaluations' => $evaluationRepository->findAll(),
+            ]);
+        } else {
+            $response = $this->redirectToRoute('security_logout');
+            $response->setSharedMaxAge(0);
+            $response->headers->addCacheControlDirective('no-cache', true);
+            $response->headers->addCacheControlDirective('no-store', true);
+            $response->headers->addCacheControlDirective('must-revalidate', true);
+            $response->setCache([
+                'max_age' => 0,
+                'private' => true,
+            ]);
+            return $response;
+        }
     }
 
     /**
@@ -33,55 +46,68 @@ class EvaluationController extends AbstractController
      */
     public function new(Request $request): Response
     {
-        $criteres = $this->getDoctrine()->getRepository(CritereEvaluation::class)->findAll();
-        $employes = $this->getDoctrine()->getRepository(Employe::class)->findAll();
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_RH')) {
+            $criteres = $this->getDoctrine()->getRepository(CritereEvaluation::class)->findAll();
+            $employes = $this->getDoctrine()->getRepository(Employe::class)->findAll();
 
-        $evaluation = new Evaluation();
-        $form = $this->createForm(EvaluationType::class, $evaluation);
-        $form->handleRequest($request);
-        if ($request->isMethod('POST')) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $employeSelect = $request->request->get('employe');
-            $dateEvaluation = new \DateTime($request->request->get('dateEvaluation'));
-            $notes = $request->request->get('notes');
-            $commentaires = $request->request->get('commentaires');
-            $employe = $entityManager->getRepository(Employe::class)->find($employeSelect);
+            $evaluation = new Evaluation();
+            $form = $this->createForm(EvaluationType::class, $evaluation);
+            $form->handleRequest($request);
+            if ($request->isMethod('POST')) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $employeSelect = $request->request->get('employe');
+                $dateEvaluation = new \DateTime($request->request->get('dateEvaluation'));
+                $notes = $request->request->get('notes');
+                $commentaires = $request->request->get('commentaires');
+                $employe = $entityManager->getRepository(Employe::class)->find($employeSelect);
 
-            $evaluation->setDateEvaluation($dateEvaluation);
-            $evaluation->setEmploye($employe);
-          
-            $totalNotes = 0;
-            $nbNotes = 0;
+                $evaluation->setDateEvaluation($dateEvaluation);
+                $evaluation->setEmploye($employe);
 
-            foreach ($notes as $critereId => $note) {
-                $critere = $entityManager->getRepository(CritereEvaluation::class)->find($critereId);
-                $evaluationDetail = new EvaluationDetail();
-                $evaluationDetail->setCritereEvaluation($critere);
-                $evaluationDetail->setNote((int)$note);
-                $evaluationDetail->setEvaluation($evaluation);
-                if (isset($commentaires[$critereId])) {
-                    $evaluationDetail->setCommentaire($commentaires[$critereId]);
+                $totalNotes = 0;
+                $nbNotes = 0;
+
+                foreach ($notes as $critereId => $note) {
+                    $critere = $entityManager->getRepository(CritereEvaluation::class)->find($critereId);
+                    $evaluationDetail = new EvaluationDetail();
+                    $evaluationDetail->setCritereEvaluation($critere);
+                    $evaluationDetail->setNote((int)$note);
+                    $evaluationDetail->setEvaluation($evaluation);
+                    if (isset($commentaires[$critereId])) {
+                        $evaluationDetail->setCommentaire($commentaires[$critereId]);
+                    }
+                    $evaluation->addEvaluationDetail($evaluationDetail);
+                    $entityManager->persist($evaluationDetail);
+
+                    $totalNotes += (int)$note;
+                    $nbNotes++;
                 }
+
+                $moyenne = ($nbNotes > 0) ? ($totalNotes / $nbNotes) : 0;
+                $evaluation->setMoyenne($moyenne);
                 $evaluation->addEvaluationDetail($evaluationDetail);
-                $entityManager->persist($evaluationDetail);
+                $entityManager->persist($evaluation);
+                $entityManager->flush();
 
-                $totalNotes+=(int)$note;
-                $nbNotes ++;
+                return $this->redirectToRoute('evaluation_index', [], Response::HTTP_SEE_OTHER);
             }
-
-            $moyenne = ($nbNotes > 0) ? ($totalNotes / $nbNotes): 0;
-            $evaluation->setMoyenne($moyenne);
-            $evaluation->addEvaluationDetail($evaluationDetail);
-            $entityManager->persist($evaluation);
-            $entityManager->flush();
-
-            return $this->redirectToRoute('evaluation_index', [], Response::HTTP_SEE_OTHER);
+            return $this->render('evaluation/evaluation.html.twig', [
+                'employes' => $employes,
+                'form' => $form->createView(),
+                'criteres' => $criteres,
+            ]);
+        } else {
+            $response = $this->redirectToRoute('security_logout');
+            $response->setSharedMaxAge(0);
+            $response->headers->addCacheControlDirective('no-cache', true);
+            $response->headers->addCacheControlDirective('no-store', true);
+            $response->headers->addCacheControlDirective('must-revalidate', true);
+            $response->setCache([
+                'max_age' => 0,
+                'private' => true,
+            ]);
+            return $response;
         }
-        return $this->render('evaluation/evaluation.html.twig', [
-            'employes' => $employes,
-            'form' => $form->createView(),
-            'criteres' => $criteres,
-        ]);
     }
 
     /**
@@ -89,9 +115,22 @@ class EvaluationController extends AbstractController
      */
     public function show(Evaluation $evaluation): Response
     {
-        return $this->render('evaluation/show.html.twig', [
-            'evaluation' => $evaluation,
-        ]);
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_RH')) {
+            return $this->render('evaluation/show.html.twig', [
+                'evaluation' => $evaluation,
+            ]);
+        } else {
+            $response = $this->redirectToRoute('security_logout');
+            $response->setSharedMaxAge(0);
+            $response->headers->addCacheControlDirective('no-cache', true);
+            $response->headers->addCacheControlDirective('no-store', true);
+            $response->headers->addCacheControlDirective('must-revalidate', true);
+            $response->setCache([
+                'max_age' => 0,
+                'private' => true,
+            ]);
+            return $response;
+        }
     }
 
     /**
@@ -99,19 +138,32 @@ class EvaluationController extends AbstractController
      */
     public function edit(Request $request, Evaluation $evaluation): Response
     {
-        $form = $this->createForm(EvaluationType::class, $evaluation);
-        $form->handleRequest($request);
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_RH')) {
+            $form = $this->createForm(EvaluationType::class, $evaluation);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('evaluation_index', [], Response::HTTP_SEE_OTHER);
+                return $this->redirectToRoute('evaluation_index', [], Response::HTTP_SEE_OTHER);
+            }
+
+            return $this->render('evaluation/edit.html.twig', [
+                'evaluation' => $evaluation,
+                'form' => $form->createView(),
+            ]);
+        } else {
+            $response = $this->redirectToRoute('security_logout');
+            $response->setSharedMaxAge(0);
+            $response->headers->addCacheControlDirective('no-cache', true);
+            $response->headers->addCacheControlDirective('no-store', true);
+            $response->headers->addCacheControlDirective('must-revalidate', true);
+            $response->setCache([
+                'max_age' => 0,
+                'private' => true,
+            ]);
+            return $response;
         }
-
-        return $this->render('evaluation/edit.html.twig', [
-            'evaluation' => $evaluation,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -119,12 +171,25 @@ class EvaluationController extends AbstractController
      */
     public function delete(Request $request, Evaluation $evaluation): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $evaluation->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($evaluation);
-            $entityManager->flush();
-        }
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_RH')) {
+            if ($this->isCsrfTokenValid('delete' . $evaluation->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($evaluation);
+                $entityManager->flush();
+            }
 
-        return $this->redirectToRoute('evaluation_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('evaluation_index', [], Response::HTTP_SEE_OTHER);
+        } else {
+            $response = $this->redirectToRoute('security_logout');
+            $response->setSharedMaxAge(0);
+            $response->headers->addCacheControlDirective('no-cache', true);
+            $response->headers->addCacheControlDirective('no-store', true);
+            $response->headers->addCacheControlDirective('must-revalidate', true);
+            $response->setCache([
+                'max_age' => 0,
+                'private' => true,
+            ]);
+            return $response;
+        }
     }
 }
