+++ src/Controller/FeedbackController.php
@@ -21,9 +21,22 @@ class FeedbackController extends AbstractController
      */
     public function index(FeedbackRepository $feedbackRepository): Response
     {
-        return $this->render('feedback/admin/index.html.twig', [
-            'feedback' => $feedbackRepository->findAll(),
-        ]);
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_EMPLOYER')) {
+            return $this->render('feedback/admin/index.html.twig', [
+                'feedback' => $feedbackRepository->findAll(),
+            ]);
+        } else {
+            $response = $this->redirectToRoute('security_logout');
+            $response->setSharedMaxAge(0);
+            $response->headers->addCacheControlDirective('no-cache', true);
+            $response->headers->addCacheControlDirective('no-store', true);
+            $response->headers->addCacheControlDirective('must-revalidate', true);
+            $response->setCache([
+                'max_age' => 0,
+                'private' => true,
+            ]);
+            return $response;
+        }
     }
 
     /**
@@ -31,26 +44,39 @@ class FeedbackController extends AbstractController
      */
     public function new(Request $request, Security $security): Response
     {
-        $feedback = new Feedback();
-        $form = $this->createForm(FeedbackType::class, $feedback);
-        $form->handleRequest($request);
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_EMPLOYER')) {
+            $feedback = new Feedback();
+            $form = $this->createForm(FeedbackType::class, $feedback);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $employe = $security->getUser();
-            $feedback->setCreatedAt(new \DateTime());
-            $feedback->setEmploye($employe);
+            if ($form->isSubmitted() && $form->isValid()) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $employe = $security->getUser();
+                $feedback->setCreatedAt(new \DateTime());
+                $feedback->setEmploye($employe);
 
-            $entityManager->persist($feedback);
-            $entityManager->flush();
+                $entityManager->persist($feedback);
+                $entityManager->flush();
 
-            return $this->redirectToRoute('feedback_index', [], Response::HTTP_SEE_OTHER);
+                return $this->redirectToRoute('feedback_index', [], Response::HTTP_SEE_OTHER);
+            }
+
+            return $this->render('feedback/admin/new.html.twig', [
+                'feedback' => $feedback,
+                'form' => $form->createView(),
+            ]);
+        } else {
+            $response = $this->redirectToRoute('security_logout');
+            $response->setSharedMaxAge(0);
+            $response->headers->addCacheControlDirective('no-cache', true);
+            $response->headers->addCacheControlDirective('no-store', true);
+            $response->headers->addCacheControlDirective('must-revalidate', true);
+            $response->setCache([
+                'max_age' => 0,
+                'private' => true,
+            ]);
+            return $response;
         }
-
-        return $this->render('feedback/admin/new.html.twig', [
-            'feedback' => $feedback,
-            'form' => $form->createView(),
-        ]);
     }
 
 
@@ -59,12 +85,25 @@ class FeedbackController extends AbstractController
      */
     public function suivi(Security $security): Response
     {
-        $employe = $security->getUser();
-        $feedback = $this->getDoctrine()->getRepository(Feedback::class)->findBy(['employe' => $employe]);
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_EMPLOYER')) {
+            $employe = $security->getUser();
+            $feedback = $this->getDoctrine()->getRepository(Feedback::class)->findBy(['employe' => $employe]);
 
-        return $this->render("feedback/index.html.twig", [
-            'feedback' => $feedback,
-        ]);
+            return $this->render("feedback/index.html.twig", [
+                'feedback' => $feedback,
+            ]);
+        } else {
+            $response = $this->redirectToRoute('security_logout');
+            $response->setSharedMaxAge(0);
+            $response->headers->addCacheControlDirective('no-cache', true);
+            $response->headers->addCacheControlDirective('no-store', true);
+            $response->headers->addCacheControlDirective('must-revalidate', true);
+            $response->setCache([
+                'max_age' => 0,
+                'private' => true,
+            ]);
+            return $response;
+        }
     }
 
     /**
@@ -72,9 +111,22 @@ class FeedbackController extends AbstractController
      */
     public function suiviShow(Feedback $feedback): Response
     {
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_EMPLOYER')) {
             return $this->render("feedback/show.html.twig", [
-            'feedback' => $feedback,
-        ]);
+                'feedback' => $feedback,
+            ]);
+        } else {
+            $response = $this->redirectToRoute('security_logout');
+            $response->setSharedMaxAge(0);
+            $response->headers->addCacheControlDirective('no-cache', true);
+            $response->headers->addCacheControlDirective('no-store', true);
+            $response->headers->addCacheControlDirective('must-revalidate', true);
+            $response->setCache([
+                'max_age' => 0,
+                'private' => true,
+            ]);
+            return $response;
+        }
     }
 
 
@@ -83,9 +135,22 @@ class FeedbackController extends AbstractController
      */
     public function show(Feedback $feedback): Response
     {
-        return $this->render('feedback/admin/show.html.twig', [
-            'feedback' => $feedback,
-        ]);
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_EMPLOYER')) {
+            return $this->render('feedback/admin/show.html.twig', [
+                'feedback' => $feedback,
+            ]);
+        } else {
+            $response = $this->redirectToRoute('security_logout');
+            $response->setSharedMaxAge(0);
+            $response->headers->addCacheControlDirective('no-cache', true);
+            $response->headers->addCacheControlDirective('no-store', true);
+            $response->headers->addCacheControlDirective('must-revalidate', true);
+            $response->setCache([
+                'max_age' => 0,
+                'private' => true,
+            ]);
+            return $response;
+        }
     }
 
     /**
@@ -93,19 +158,32 @@ class FeedbackController extends AbstractController
      */
     public function edit(Request $request, Feedback $feedback): Response
     {
-        $form = $this->createForm(FeedbackType::class, $feedback);
-        $form->handleRequest($request);
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_EMPLOYER')) {
+            $form = $this->createForm(FeedbackType::class, $feedback);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('feedback_index', [], Response::HTTP_SEE_OTHER);
+                return $this->redirectToRoute('feedback_index', [], Response::HTTP_SEE_OTHER);
+            }
+
+            return $this->render('feedback/admin/edit.html.twig', [
+                'feedback' => $feedback,
+                'form' => $form->createView(),
+            ]);
+        } else {
+            $response = $this->redirectToRoute('security_logout');
+            $response->setSharedMaxAge(0);
+            $response->headers->addCacheControlDirective('no-cache', true);
+            $response->headers->addCacheControlDirective('no-store', true);
+            $response->headers->addCacheControlDirective('must-revalidate', true);
+            $response->setCache([
+                'max_age' => 0,
+                'private' => true,
+            ]);
+            return $response;
         }
-
-        return $this->render('feedback/admin/edit.html.twig', [
-            'feedback' => $feedback,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -113,12 +191,25 @@ class FeedbackController extends AbstractController
      */
     public function delete(Request $request, Feedback $feedback): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $feedback->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($feedback);
-            $entityManager->flush();
-        }
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_EMPLOYER')) {
+            if ($this->isCsrfTokenValid('delete' . $feedback->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($feedback);
+                $entityManager->flush();
+            }
 
-        return $this->redirectToRoute('feedback_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('feedback_index', [], Response::HTTP_SEE_OTHER);
+        } else {
+            $response = $this->redirectToRoute('security_logout');
+            $response->setSharedMaxAge(0);
+            $response->headers->addCacheControlDirective('no-cache', true);
+            $response->headers->addCacheControlDirective('no-store', true);
+            $response->headers->addCacheControlDirective('must-revalidate', true);
+            $response->setCache([
+                'max_age' => 0,
+                'private' => true,
+            ]);
+            return $response;
+        }
     }
 }
